# docs/openapi/authorization-service.yaml
openapi: 3.0.3
info: {title: Aurevia Internal Authorization Service, version: 0.2.0}
security: [{workloadIdentity: []}]
paths:
...
  /internal/v1/registry/grants:
    post:
      operationId: createGrant
      parameters: [{$ref: '#/components/parameters/Actor'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GrantWrite'}}}}
      responses: {'201': {description: Grant and outbox event created}, '403': {description: Admin permission required}, '409': {description: Active grant exists}}
  /internal/v1/registry/resource-tree:
    get: {operationId: resourceTree, parameters: [{$ref: '#/components/parameters/Actor'}], responses: {'200': {description: Complete seven-type resource catalog with actions and grant counts}}}
  /internal/v1/registry/resource-types:
    get: {operationId: resourceTypes, parameters: [{$ref: '#/components/parameters/Actor'}], responses: {'200': {description: Supported resource type names}}}
  /internal/v1/registry/subjects/{type}/{id}/grants:
    get:
      operationId: subjectGrants
      parameters:
        - {$ref: '#/components/parameters/Actor'}
        - {in: path, name: type, required: true, schema: {type: string, enum: [USER, GROUP, ROLE]}}
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      responses: {'200': {description: Active direct grants for the selected subject}}
  /internal/v1/registry/grants/{id}:
    delete:
      operationId: revokeGrant
      parameters:
        - {$ref: '#/components/parameters/Actor'}
        - {in: path, name: id, required: true, schema: {type: string, format: uuid}}
      responses: {'204': {description: Grant archived and delete event enqueued}}
...
components:
  securitySchemes:
    workloadIdentity: {type: http, scheme: basic, description: Local bootstrap; production requires mTLS.}
  parameters:
    Actor: {in: header, name: X-Actor, required: true, schema: {type: string}}
  schemas:
...
    GrantWrite:
      type: object
      required: [resourceId, actionId, relation]
      properties:
        userId: {type: string, format: uuid, deprecated: true}
        subjectType: {type: string, enum: [USER, GROUP, ROLE]}
        subjectId: {type: string, format: uuid}
        resourceId: {type: string, format: uuid}
        actionId: {type: string, format: uuid}
        relation: {type: string}
        expiresAt: {type: string, format: date-time, nullable: true}